remote_nodes/landet_node/sensor_logger/requirements.txt
paho_mqtt
asyncio-mqtt
python-binary-memcached
aiofiles
aiohttp
ujson
aiotinydb
